+++ modules/core/import/modules/csv/tests/src/Kernel/CsvImportTestBase.php
@@ -51,7 +51,6 @@ class CsvImportTestBase extends MigrateTestBase {
     'migrate',
     'migrate_plus',
     'migrate_source_csv',
-    'migrate_source_ui',
     'migrate_tools',
     'options',
     'quantity',